modularize dashboard
