+++ src/App.css
@@ -65,7 +65,6 @@
 }
 
 
-
 .dashboard-page {
     display: flex;
     flex-direction: column;
